requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=./requirements.txt ./requirements.in
#
openfeature-sdk==0.0.9
    # via -r ./requirements.in
